workspaces/MPWR_Payment_Agent/requirements.txt
playwright
supabase
python-dotenv
pytz
apscheduler
greenlet
requests
